+++ Demo/MainViewModel.cs
@@ -5,7 +5,6 @@
 // -----------------------------------------------------------------------------------------------------------------
 
 using System;
-using System.Windows.Forms.VisualStyles;
 using Chapter.Net;
 using static Demo.PropertyViewModel;
 
@@ -24,6 +23,7 @@ public class MainViewModel : ObservableObject
     public PropertyViewModel SingleCharacterCheckToVisibilityConverter { get; set; } = new('a', '0', 'c');
     public PropertyViewModel MultiCharacterCheckToVisibilityConverter { get; set; } = new('a', '0', 'c');
     public PropertyViewModel ConcatenateStringConverter { get; set; } = new("First", "Second", "Third");
+    public PropertyViewModel ValueConverterGroup { get; set; } = new(9, 9, 9);
     public PropertyViewModel DateOnlyToStringConverter { get; set; } = new(new DateOnly(DateTime.Now.Year, DateTime.Now.Month, DateTime.Now.Day));
     public PropertyViewModel DateTimeToStringConverter { get; set; } = new(DateTime.Now);
     public PropertyViewModel DateTimeUtcConverter { get; set; } = new(DateTime.Now);